EUI-1115 replaced navigation handler, unfinished
